backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# OAuth & Auth
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# Environment & Config
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Password Hashing
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Email
email-validator==2.1.0

# Utils
python-dateutil==2.8.2


